Moved boost::swap into a detail namespace to avoid it being found via ADL.
This resolves infinite recursion when boost::swap is called on types that
are defined in namespace boost. The swap_impl implementation and noexcept
specification is using an unqualified call to swap, so if the argument types
are defined in namespace boost then our boost::swap function may be found
and selected to resolve the call.

Fixes https://github.com/boostorg/core/issues/148.